Increase loc precision and improve basic plume

+++ src/projection.h
@@ -9,32 +9,32 @@ class Projection {
 public:
     // Set externally
     int code; // Projection code
-    float lat1;
-    float lon1;
-    float knowni; // X-location of known lat/lon
-    float knownj; // Y-location of known lat/lon
-    float dx;
-    float stdlon;
-    float truelat1;
-    float truelat2;
+    double lat1;
+    double lon1;
+    double knowni; // X-location of known lat/lon
+    double knownj; // Y-location of known lat/lon
+    double dx;
+    double stdlon;
+    double truelat1;
+    double truelat2;
     // Set internally
-    float dlat; // Lat increment for lat/lon grids
-    float dlon; // Lon increment for lat/lon grids
+    double dlat; // Lat increment for lat/lon grids
+    double dlon; // Lon increment for lat/lon grids
     int hemi; // 1 for NH, -1 for SH
-    float cone; // Cone factor for LC projections
-    float polei; // i-location of pole point
-    float polej; // j-location of pole point
-    float rsw; // Radius to SW corner
-    float rebydx; // Earth radius divided by dx
-    float ctl1r; // cos(truelat1*RAD_PER_DEG)
+    double cone; // Cone factor for LC projections
+    double polei; // i-location of pole point
+    double polej; // j-location of pole point
+    double rsw; // Radius to SW corner
+    double rebydx; // Earth radius divided by dx
+    double ctl1r; // cos(truelat1*RAD_PER_DEG)
     bool isInitd = false;
 
-    void init(int proj_code, float lat1, float lon1, float knowni, float knownj, float dx,
-              float stdlon, float truelat1, float truelat2);
+    void init(int proj_code, double lat1, double lon1, double knowni, double knownj, double dx,
+              double stdlon, double truelat1, double truelat2);
 
     void set_lc();
 
-    float lc_cone(float truelat1, float truelat2);
+    double lc_cone(double truelat1, double truelat2);
 
     IDX2 loc_to_ij(const Geo2D& loc) const;
